requirements.txt
flask
mysql-connector-python
opencv-python
opencv-contrib-python
numpy
